pyproject.toml
[project]
name = "bio-info"
version = "0.1.0"
description = "A tool to fetch research papers with non-academic authors from PubMed"
authors = [{ name = "Vivek", email = "viveksinghcharak004@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "typing-extensions (>=4.12.2,<5.0.0)",
    "biopython (>=1.85,<2.0)",
]
license = "Proprietary"

[tool.poetry]
packages = [{ include = "bio_info", from = "src" }]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
get-papers-list = "bio_info.cli:main"
